requirements.txt
# Hackathon0Gold Python Dependencies
# Install with: pip install -r requirements.txt

# Browser Automation & Web Scraping
playwright==1.48.0
selenium==4.27.1
undetected-chromedriver==3.5.4
webdriver-manager==4.0.2

# Google API & Authentication
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==1.12.5

# Process Management (optional, PM2 is Node.js based)
pm2-python-wrapper==1.0.0

# Data & File Operations
python-dotenv==1.0.1
pyyaml==6.0.2
toml==0.10.2

# Logging & Monitoring
colorama==0.4.6
tqdm==4.68.0

# Async Operations
aiohttp==3.10.11
asyncio-contextmanager==1.0.0

# API & HTTP
requests==2.32.3
httpx==0.28.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development & Testing (optional)
pytest==8.3.4
pytest-asyncio==0.24.0
black==25.1.1
pylint==3.3.1
